Enhance dtype handling and inference in pipeline operations
- Updated the dtype tracking in the Pipeline class to start as "auto", allowing for better handling of unknown types until runtime.
- Improved the DTypeEffect class to resolve output types based on input conditions, particularly for image sources.
- Refined validation logic for list and array sinks to ensure known dtypes are enforced at planning time.
- Adjusted tests to reflect changes in dtype propagation, ensuring consistency across various operations and sources.
- Enhanced Rust backend to support new dtype inference logic, improving overall robustness and accuracy in data processing.

+++ polars-cv/tests/test_serialization.py
@@ -108,7 +108,7 @@ class TestPipelineJsonFormat:
 
     def test_sink_spec_array_shape(self) -> None:
         """Test sink specification with array shape."""
-        pipe = Pipeline().source().sink("array", shape=[224, 224, 3])
+        pipe = Pipeline().source(dtype="u8").sink("array", shape=[224, 224, 3])
         data = json.loads(pipe._to_json())
 
         assert data["sink"]["format"] == "array"
